+++ sql-plugin/src/main/spark320/scala/com/nvidia/spark/rapids/shims/GetSequenceSize.scala
@@ -39,15 +39,7 @@ package com.nvidia.spark.rapids.shims
 import ai.rapids.cudf._
 import com.nvidia.spark.rapids.Arm._
 
-import org.apache.spark.unsafe.array.ByteArrayMethods.MAX_ROUNDED_ARRAY_LENGTH
-
 object GetSequenceSize {
-  def TOO_LONG_SEQUENCE(sequenceLength: Int, functionName: String) = {
-    // For these Spark versions, the sequence length and function name
-    // do not appear in the exception message.
-    s"Too long sequence found. Should be <= $MAX_ROUNDED_ARRAY_LENGTH"
-  }
-
   /**
    * Compute the size of each sequence according to 'start', 'stop' and 'step'.
    * A row (Row[start, stop, step]) contains at least one null element will produce
